Controllers Fused
studentEdit, AddSingleStudent and SwitchStudent fused into one controller so that we have less files to worry about
