+++ htdocs/app/Config/Routes.php
@@ -58,26 +58,19 @@ $routes->match(['get', 'post'], 'ClassCreateController/CreateClass', 'ClassCreat
 $routes->get('/createModPage', 'AdminController::createModPage');
 $routes->post('/createModAccount', 'AdminController::createModerator');
 
-$routes->get('/AddSingleStudent/(:alphanum)','AddSingleStudentController::index/$1');
-$routes->match(['get', 'post'], 'AddSingleStudentController/CreateUser', 'AddSingleStudentController::CreateUser');
+$routes->get('/AddSingleStudent/(:alphanum)','AllStudentFeaturesController::indexAddSingleStudent/$1');
+$routes->match(['get', 'post'], 'AddSingleStudentController/CreateUser', 'AllStudentFeaturesController::CreateUser');
 
 $routes->get('/DeleteClass/(:alphanum)','DeleteClassController::Delete/$1');
 
 $routes->get('/deleteModerator/(:alphanum)', 'AdminController::deleteModerator/$1');
-
 
 $routes->get('/classes/(:alphanum)','ClassViewController::index/$1');
-$routes->get('/StudentEdit/(:alphanum)','StudentEditController::index/$1');
-$routes->match(['get', 'post'], 'StudentEditController/UpdateUser', 'StudentEditController::CreateClass');
-
-
+$routes->get('/StudentEdit/(:alphanum)','AllStudentFeaturesController::indexStudentEdit/$1');
+$routes->match(['get', 'post'], 'AllStudentFeaturesController/UpdatenUsers', 'AllStudentFeaturesController::UpdatenUsers');
 
 $routes->get('/editModerator/(:alphanum)', 'AdminController::EditModeratorPage/$1');
 $routes->match(['get', 'post'], '/updateModAccount', 'AdminController::updateModAccount');
-
-$routes->get('/SwitchClass/(:alphanum)','SwitchClassController::index/$1');
-$routes->match(['get', 'post'], 'SwitchClassController/SwitchStudent', 'SwitchClassController::SwitchStudent');
-
 
 $routes->get('/AddSingleStudent/(:alphanum)','AddSingleStudentController::index/$1');
 $routes->match(['get', 'post'], 'AddSingleStudentController/CreateUser', 'AddSingleStudentController::CreateUser');
@@ -89,9 +82,8 @@ $routes->get('/deleteModerator/(:alphanum)', 'AdminController::deleteModerator/$
 $routes->get('/editModerator/(:alphanum)', 'AdminController::EditModeratorPage/$1');
 $routes->match(['get', 'post'], '/updateModAccount', 'AdminController::updateModAccount');
 
-$routes->get('/SwitchClass/(:alphanum)','SwitchClassController::index/$1');
-$routes->match(['get', 'post'], 'SwitchClassController/SwitchStudent', 'SwitchClassController::SwitchStudent');
-
+$routes->get('/SwitchClass/(:alphanum)','AllStudentFeaturesController::indexSwitchStudent/$1');
+$routes->match(['get', 'post'], 'SwitchClassController/SwitchStudent', 'AllStudentFeaturesController::SwitchStudent');
 
 $routes->get('/ClassesEdit/(:alphanum)', 'UpdateClassController::index/$1');
 $routes->match(['get', 'post'], 'UpdateClassController/UpdateClass', 'UpdateClassController::UpdateClass');
