+++ htdocs/app/Views/homepages/moderator/StudentEdit.php
@@ -1,7 +1,7 @@
 <div class="block">
     <h3>Voeg student toe</h3>
     <br>
-    <form action="<?php echo base_url(); ?>/StudentEditController/UpdatenUsers" method="post" class="form form--rounded">
+    <form action="<?php echo base_url(); ?>/AllStudentFeaturesController/UpdatenUsers" method="post" class="form form--rounded">
         <div>
             <input type="text" name="name" value="<?php echo $HoldID["Name"] ?>"></input>
         </div>
